# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Comic API",
        "description": "AdamZ Easy Comic API",
        "contact": {
            "email": "N00191010@iadt.ie"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0.0"
    },
    "paths": {
        "/api/authors": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Display a listing of the resource.",
                "description": "Displays all the authors",
                "operationId": "0c71439e213e1b284fc37995d74bca2b",
                "responses": {
                    "200": {
                        "description": "Task failed unsuccessfully, Returns a list of all the authors in JSON format"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/author": {
            "post": {
                "tags": [
                    "Authors"
                ],
                "summary": "Create a new Author",
                "description": "Stores the author in the DB",
                "operationId": "store author",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "genre",
                                    "author",
                                    "illustrator",
                                    "issues"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Sample Name"
                                    },
                                    "address": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Address of the Author"
                                    },
                                    "bio": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "The biography of the author"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/authors/{id}": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Displays the specified author by said ID.",
                "description": "Gets a authors by ID",
                "operationId": "c14c739d68f46a33acfd0694a12c27dc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Author id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful task"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/author/{id}": {
            "put": {
                "tags": [
                    "Authors"
                ],
                "summary": "Update the specified resource in storage.",
                "description": "Updates a Author by ID",
                "operationId": "95f3dc321fe15b0c7f6f4984f80c9b9c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Author id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful task"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Authors"
                ],
                "summary": "Delete a Author",
                "description": "Delete Author",
                "operationId": "delete author",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of a Author",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "204"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/comics": {
            "get": {
                "tags": [
                    "Comics"
                ],
                "summary": "Display a listing of the resource.",
                "description": "Displays all the comics",
                "operationId": "98dc4ed0037dc558691bd498614669ec",
                "responses": {
                    "200": {
                        "description": "Task failed unsuccessfully, Returns a list of Comics in a JSON format"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Comics"
                ],
                "summary": "Create a new Comic",
                "description": "Stores the comic in the DB",
                "operationId": "store comic",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "genre",
                                    "author",
                                    "illustrator",
                                    "issues"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Sample Title"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Interesting bit of text"
                                    },
                                    "genre": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "The category of literature"
                                    },
                                    "author": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Person"
                                    },
                                    "illustrator": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Person"
                                    },
                                    "issues": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/comics/{id}": {
            "get": {
                "tags": [
                    "Comics"
                ],
                "summary": "Displays the specified comic by said ID.",
                "description": "Gets a comic by ID",
                "operationId": "085385974afa441ba0353a39920c391f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Comic id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful task"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Comics"
                ],
                "summary": "Update the specified resource in storage.",
                "description": "Updates a comic by ID",
                "operationId": "44f276d6148d41d33a52681a8d27579e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Comic id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful task"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/comic/{id}": {
            "delete": {
                "tags": [
                    "Comics"
                ],
                "summary": "Delete a Comic",
                "description": "Delete Comic",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of a Comic",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "204"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/distributor": {
            "get": {
                "tags": [
                    "Distributors"
                ],
                "summary": "Display a listing of the resource.",
                "description": "Displays all the distributors",
                "operationId": "91b717ff23cef497db5da7bc474aa3c8",
                "responses": {
                    "200": {
                        "description": "Task failed unsuccessfully, Returns a list of distributors in a JSON format"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Distributors"
                ],
                "summary": "Create a new Distributor",
                "description": "Stores the distributor in the DB",
                "operationId": "store distributor",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "biography"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Sample Name"
                                    },
                                    "biography": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Bit of text for the biography"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/distributors/{id}": {
            "get": {
                "tags": [
                    "Distributors"
                ],
                "summary": "Displays the specified distributor by said ID.",
                "description": "Gets a distributors by ID",
                "operationId": "d67e88a587be2a4b46bf8ccab6d243bd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Distributor id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful task"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/distributor/{id}": {
            "put": {
                "tags": [
                    "Distributors"
                ],
                "summary": "Update the specified resource in storage.",
                "description": "Updates a distributor by ID",
                "operationId": "26d1327a3ee739cc6667711e31e10d27",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Comic id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful task"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Distributors"
                ],
                "summary": "Delete a Distributor",
                "description": "Delete Distributor",
                "operationId": "delete ",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of a Distributor",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "204"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}